把工時改為 float
--資料庫調整
ALTER TABLE `todo` CHANGE `hours` `hours` FLOAT( 11 ) NOT NULL COMMENT '預估時間';
ALTER TABLE `todo` CHANGE `usedHours` `usedHours` FLOAT( 11 ) NOT NULL ;

+++ application/controllers/page/show.php
@@ -22,9 +22,9 @@ class Show extends CI_Controller
 		{ 
             $data[$i]->note         = $this->_note($data[$i]->note);
             $data[$i]->createTime   = $this->_time($data[$i]->createTime);
-            $data[$i]->hours        = ($data[$i]->hours) ? intval($data[$i]->hours) : '-';
-            $data[$i]->usedHours    = ($data[$i]->usedHours) ? intval($data[$i]->usedHours) : '-';
-            $data[$i]->surplusHours = ($data[$i]->hours != '-') ? $data[$i]->hours - intval($data[$i]->usedHours) : '-';
+            $data[$i]->hours        = ($data[$i]->hours) ? floatval($data[$i]->hours) : '-';
+            $data[$i]->usedHours    = ($data[$i]->usedHours) ? floatval($data[$i]->usedHours) : '-';
+            $data[$i]->surplusHours = ($data[$i]->hours != '-') ? $data[$i]->hours - floatval($data[$i]->usedHours) : '-';
 		}
 
 		// 整理 view data
